requirements.txt
crewai
langchain_groq
python-dotenv
groq
litellm
streamlit